refactor(uk-fuel-finder): split token + parser out of uk_fuel_finder_service (Refs #563)
- uk_fuel_finder_service.dart: 380 -> 191 LOC (service shell only).
- lib/core/services/impl/uk_fuel_finder_token_manager.dart (new): OAuth2 client-credentials token lifecycle.
- lib/core/services/impl/uk_fuel_finder_response_parser.dart (new): JSON-to-Station parsing, pence-to-EUR conversion.
- UkFuelFinderService keeps public test surface: kClientIdStorageKey, kClientSecretStorageKey, parsePenceForTest, parseFuelFinderStations, expireTokenForTest. Tests unchanged.
